change udacity-c3-restapi to have two services
Divide this backend server into two services: user and feed
