+++ lesson-1-Best-Practices/exercises/udacity-c3-restapi-feed/src/controllers/v0/feed/routes/feed.router.ts
@@ -1,9 +1,33 @@
 import { Router, Request, Response } from 'express';
 import { FeedItem } from '../models/FeedItem';
-import { requireAuth } from '../../users/routes/auth.router';
+import * as jwt from 'jsonwebtoken';
+import * as c from '../../../../config/config';
+//import { requireAuth } from '../../users/routes/auth.router';
 import * as AWS from '../../../../aws';
 
 const router: Router = Router();
+
+
+export function requireAuth(req: Request, res: Response, next: NextFunction) {
+    //  return next();
+       if (!req.headers || !req.headers.authorization){
+           return res.status(401).send({ message: 'No authorization headers.' });
+       }
+       
+   
+       const token_bearer = req.headers.authorization.split(' ');
+       if(token_bearer.length != 2){
+           return res.status(401).send({ message: 'Malformed token.' });
+       }
+       
+       const token = token_bearer[1];
+       return jwt.verify(token, c.config.jwt.secret , (err, decoded) => {
+         if (err) {
+           return res.status(500).send({ auth: false, message: 'Failed to authenticate.' });
+         }
+         return next();
+       });
+   }
 
 // Get all feed items
 router.get('/', async (req: Request, res: Response) => {
