+++ lesson-1-Best-Practices/exercises/udacity-c3-restapi-feed/src/controllers/v0/index.router.ts
@@ -1,11 +1,11 @@
 import { Router, Request, Response } from 'express';
 import { FeedRouter } from './feed/routes/feed.router';
-import { UserRouter } from './users/routes/user.router';
+//import { UserRouter } from './users/routes/user.router';
 
 const router: Router = Router();
 
 router.use('/feed', FeedRouter);
-router.use('/users', UserRouter);
+//router.use('/users', UserRouter);
 
 router.get('/', async (req: Request, res: Response) => {    
     res.send(`V0`);
