+++ lesson-1-Best-Practices/exercises/udacity-c3-restapi-user/src/controllers/v0/index.router.ts
@@ -1,10 +1,10 @@
 import { Router, Request, Response } from 'express';
-import { FeedRouter } from './feed/routes/feed.router';
+// import { FeedRouter } from './feed/routes/feed.router';
 import { UserRouter } from './users/routes/user.router';
 
 const router: Router = Router();
 
-router.use('/feed', FeedRouter);
+// router.use('/feed', FeedRouter);
 router.use('/users', UserRouter);
 
 router.get('/', async (req: Request, res: Response) => {
