+++ lesson-1-Best-Practices/exercises/udacity-c3-restapi-user/src/controllers/v0/model.index.ts
@@ -1,4 +1,5 @@
-import { FeedItem } from './feed/models/FeedItem';
+//import { FeedItem } from './feed/models/FeedItem';
 import { User } from './users/models/User';
 
-export const V0MODELS = [ FeedItem, User ];
+//export const V0MODELS = [ FeedItem, User ];
+export const V0MODELS = [ User ];
